requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.3
Django==6.0
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-stubs==5.2.8
django-stubs-ext==5.2.8
soupsieve==2.8
sqlparse==0.5.4
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
webvtt-py==0.5.1
